+++ src/app/beer_garden/db/mongo/models.py
@@ -509,8 +509,17 @@ class Request(MongoModel, Document):
             )
 
     def _post_save(self):
-        if self.status == "CREATED" and self.namespace == config.get("garden.name"):
-            self._update_raw_file_references()
+        if self.status == "CREATED":
+            try:
+                ref_system = System.objects.get(
+                    namespace=self.namespace,
+                    name=self.system,
+                    version=self.system_version,
+                )
+                if ref_system.local:
+                    self._update_raw_file_references()
+            except self.DoesNotExist:
+                pass
 
     def _update_raw_file_references(self):
         parameters = self.parameters or {}