+++ src/app/beer_garden/router.py
@@ -742,13 +742,7 @@ def _pre_forward(operation: Operation) -> Operation:
         # Save the request so it'll have an ID and we'll have something to update
         local_request = create_request(operation.model)
 
-        if operation.model.namespace == config.get("garden.name"):
-            operation.model = local_request
-        else:
-            # When the target is a remote garden, just capture the id. We don't
-            # want to replace the entire model, as we'd lose the base64 encoded file
-            # parameter data.
-            operation.model.id = local_request.id
+        operation.model.id = local_request.id
 
         beer_garden.files.forward_file(operation)